libxcb mesa pthreads git update April 4 2017
libxcb           commit f830eb93c9c38b2c6c7ea2971af3bc6a61e92277
mesa             commit 7216e1d8afedda9dd0f8a6ab0e03274190d1c7be
pthreads         commit 9bee0860555dc00e8fc3b6fcac71870899c1a512

+++ mesalib/src/util/u_vector.h
@@ -38,7 +38,7 @@
 static inline uint32_t
 u_align_u32(uint32_t v, uint32_t a)
 {
-   assert(a != 0 && a == (a & -a));
+   assert(a != 0 && a == (a & -((int32_t) a)));
    return (v + a - 1) & ~(a - 1);
 }
